# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  domain.Leave:
    properties:
      created_at:
        example: "2024-01-01T00:00:00Z"
        type: string
      employee_id:
        example: 123
        type: integer
      from_date:
        example: "2024-01-15T00:00:00Z"
        type: string
      id:
        example: 1
        type: integer
      reason:
        example: Family vacation
        type: string
      status:
        allOf:
        - $ref: '#/definitions/domain.LeaveStatus'
        example: pending
      to_date:
        example: "2024-01-20T00:00:00Z"
        type: string
    type: object
  domain.LeaveStatus:
    enum:
    - pending
    - approved
    - rejected
    type: string
    x-enum-comments:
      Approved: Approved represents an approved leave request
      Pending: Pending represents a pending leave request
      Rejected: Rejected represents a rejected leave request
    x-enum-descriptions:
    - Pending represents a pending leave request
    - Approved represents an approved leave request
    - Rejected represents a rejected leave request
    x-enum-varnames:
    - Pending
    - Approved
    - Rejected
  handler.ErrorResponse:
    properties:
      error:
        example: Invalid request body
        type: string
    type: object
  handler.RegisterResponse:
    properties:
      message:
        example: User registered successfully
        type: string
    type: object
  handler.registerRequest:
    properties:
      email:
        example: john@example.com
        type: string
      password:
        example: password123
        minLength: 8
        type: string
      password_confirmation:
        example: password123
        minLength: 8
        type: string
      username:
        example: john_doe
        type: string
    required:
    - email
    - password
    - password_confirmation
    - username
    type: object
host: localhost:3000
info:
  contact: {}
  description: Human Resource Management Backend API with Swagger documentation
  title: HRM Backend API
  version: "1.0"
paths:
  /leaves/{id}/approve:
    put:
      description: Approve a leave request by ID
      parameters:
      - description: Leave ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
      summary: Approve leave
      tags:
      - leaves
  /leaves/{id}/reject:
    put:
      description: Reject a leave request by ID
      parameters:
      - description: Leave ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
      summary: Reject leave
      tags:
      - leaves
  /leaves/all:
    get:
      description: Get all leave requests
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Leave'
            type: array
        "500":
          description: Internal Server Error
      summary: Get all leaves
      tags:
      - leaves
  /leaves/request:
    post:
      consumes:
      - application/json
      description: Submit a new leave request
      parameters:
      - description: Leave request data
        in: body
        name: leave
        required: true
        schema:
          $ref: '#/definitions/domain.Leave'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Leave'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
      summary: Request leave
      tags:
      - leaves
  /users/register:
    post:
      consumes:
      - application/json
      description: Register a new user account
      parameters:
      - description: User registration data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/handler.registerRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handler.RegisterResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Register a new user
      tags:
      - users
swagger: "2.0"